requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
reportlab>=4.0
streamlit>=1.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
